circuits: zk-circuits: valid-match-settle: Restructure for discoverability
